+++ Source/Objects/Data Processing/DataExplorer/ORDataExplorerController.h
@@ -97,7 +97,7 @@
 
 #pragma mark ¥¥¥Data Source Methods
 - (void) tableViewSelectionDidChange:(NSNotification *)aNotification;
-- (id) tableView:(NSTableView *) aTableView objectValueForTableColumn:(NSTableColumn *) aTableColumn row:(int) rowIndex;
+- (id) tableView:(NSTableView *) aTableView objectValueForTableColumn:(NSTableColumn *) aTableColumn row:(NSInteger) rowIndex;
 - (NSInteger)numberOfRowsInTableView:(NSTableView *)aTableView;
 - (id)outlineView:(NSOutlineView *)outlineView objectValueForTableColumn:(NSTableColumn *)tableColumn byItem:(id)item; 
 - (id)outlineView:(NSOutlineView *)outlineView child:(int)index ofItem:(id)item;
